add only_changed input to run rubocop only against changed files
Switch to bash for process substitution and working with arrays.

Ignore if there are more than 100. Protecting from possibly hitting the
command line length limit, it can be much higher and can be calculated or
also extracted as one more input if needed. The effect from limiting
number of files processed by rubocop is also smaller.